# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ping App",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/devices/": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Get device list",
                "description": "Get device list",
                "operationId": "getDeviceList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DevicesListResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Store device",
                "description": "Store device",
                "operationId": "storeDevice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeviceStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Device created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceStoreResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "DeviceStoreRequest": {
                "required": [
                    "uuid",
                    "name"
                ],
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Device 1"
                    }
                },
                "type": "object"
            },
...
            "DeviceStoreResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Device 1"
                    }
                },
                "type": "object"
            },
            "DevicesListResource": {
                "properties": {
                    "uuid": {
                        "type": "string",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Device 1"
                    },
                    "last_battery_percenatge": {
                        "type": "integer",
                        "example": 100
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}